requirements.txt
pandas>=2.2
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.2
python-dateutil>=2.9
pyyaml>=6.0
